--- node/service/Cargo.toml
@@ -1,139 +1,0 @@
-[package]
-name = "tangle-service"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-clap = { workspace = true }
-futures = { workspace = true }
-hex = { workspace = true }
-hex-literal = { workspace = true }
-log = { workspace = true }
-parity-scale-codec = { workspace = true }
-rand = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-
-# Light client dependencies
-pallet-eth2-light-client-relayer-gadget = { workspace = true, optional = true }
-pallet-eth2-light-client-relayer-gadget-cli = { workspace = true, optional = true }
-webb-consensus-types = { workspace = true }
-webb-proposals = { workspace = true }
-
-tangle-relayer-gadget = { workspace = true, optional = true }
-tangle-relayer-gadget-cli = { workspace = true, optional = true }
-
-# Substrate dependencies
-frame-benchmarking = { workspace = true }
-frame-benchmarking-cli = { workspace = true }
-frame-system = { workspace = true }
-sc-authority-discovery = { workspace = true }
-sc-basic-authorship = { workspace = true }
-sc-cli = { workspace = true }
-sc-client-api = { workspace = true }
-sc-consensus = { workspace = true }
-sc-consensus-aura = { workspace = true }
-sc-consensus-babe = { workspace = true }
-sc-sysinfo = { workspace = true }
-sc-consensus-grandpa = { workspace = true }
-sc-executor = { workspace = true }
-sc-keystore = { workspace = true }
-sc-network = { workspace = true }
-sc-network-sync = { workspace = true }
-sc-offchain = { workspace = true }
-sc-service = { workspace = true }
-sc-statement-store = { workspace = true }
-sc-telemetry = { workspace = true }
-sc-transaction-pool = { workspace = true }
-sp-consensus-aura = { workspace = true }
-sp-consensus-babe = { workspace = true }
-sp-statement-store = { workspace = true }
-sp-core = { workspace = true }
-sp-inherents = { workspace = true }
-sp-keyring = { workspace = true }
-sp-runtime = { workspace = true }
-sp-trie = { workspace = true }
-
-sc-chain-spec = { workspace = true }
-sc-consensus-manual-seal = { workspace = true }
-sc-rpc = { workspace = true }
-sc-rpc-api = { workspace = true }
-sp-application-crypto = { workspace = true }
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-sp-blockchain = { workspace = true }
-sp-timestamp = { workspace = true }
-sp-io = { workspace = true }
-sp-keystore = { workspace = true }
-sp-session = { workspace = true }
-sp-offchain = { workspace = true }
-sp-transaction-pool = { workspace = true }
-sp-storage = { workspace = true }
-sp-consensus = { workspace = true }
-sp-consensus-grandpa = { workspace = true }
-substrate-frame-rpc-system = { workspace = true }
-substrate-prometheus-endpoint = { workspace = true }
-
-# RPC related dependencies
-jsonrpsee = { workspace = true }
-pallet-im-online = { workspace = true }
-pallet-transaction-payment = { workspace = true }
-pallet-transaction-payment-rpc = { workspace = true }
-pallet-transaction-payment-rpc-runtime-api = { workspace = true }
-sc-transaction-pool-api = { workspace = true }
-wasmer = { version = "2.3" }
-
-# Frontier
-fc-api = { workspace = true }
-fc-cli = { workspace = true }
-fc-consensus = { workspace = true }
-fc-db = { workspace = true }
-fc-mapping-sync = { workspace = true }
-fc-rpc = { workspace = true, features = ["rpc-binary-search-estimate"] }
-fc-rpc-core = { workspace = true }
-fc-storage = { workspace = true }
-fp-account = { workspace = true }
-fp-dynamic-fee = { workspace = true, features = ["default"] }
-fp-evm = { workspace = true, features = ["default"] }
-fp-rpc = { workspace = true, features = ["default"] }
-
-primitives-ext = { workspace = true }
-rpc-debug = { workspace = true }
-rpc-primitives-debug = { workspace = true }
-rpc-primitives-txpool = { workspace = true }
-rpc-trace = { workspace = true }
-rpc-txpool = { workspace = true }
-
-tangle-mainnet-runtime = { workspace = true }
-tangle-primitives = { workspace = true }
-tangle-testnet-runtime = { workspace = true }
-
-[features]
-default = ["with-rocksdb-weights", "rocksdb", "sql", "testnet"]
-runtime-benchmarks = ["tangle-testnet-runtime/runtime-benchmarks"]
-integration-tests = ["tangle-testnet-runtime/integration-tests"]
-with-rocksdb-weights = ["tangle-testnet-runtime/with-rocksdb-weights"]
-with-paritydb-weights = ["tangle-testnet-runtime/with-paritydb-weights"]
-rocksdb = [
-  "sc-cli/rocksdb",
-  "sc-service/rocksdb",
-  "fc-cli/rocksdb",
-  "fc-db/rocksdb",
-  "fc-mapping-sync/rocksdb",
-  "fc-rpc/rocksdb",
-]
-sql = ["fc-db/sql", "fc-mapping-sync/sql"]
-txpool = ["fc-rpc/txpool"]
-light-client = [
-  "pallet-eth2-light-client-relayer-gadget",
-  "pallet-eth2-light-client-relayer-gadget-cli",
-]
-relayer = [
-  "tangle-relayer-gadget",
-  "tangle-relayer-gadget-cli",
-]
-# Feature flags for Tangle Mainnet runtime/service
-tangle = []
-
-# Feature flags for Tangle Testnet runtime/service
-testnet = []
