--- node/service/src/client.rs
@@ -1,485 +1,0 @@
-// Copyright 2019-2022 PureStake Inc.
-// Copyright 2022 Webb Technologies Inc.
-// This file is part of Tangle.
-
-// Tangle is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Tangle is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Tangle.  If not, see <http://www.gnu.org/licenses/>.
-
-use parity_scale_codec::Encode;
-use sc_client_api::{Backend as BackendT, BlockBackend, BlockchainEvents, KeysIter, PairsIter};
-use sp_api::{CallApiAt, NumberFor, ProvideRuntimeApi};
-use sp_blockchain::{HeaderBackend, Info as BlockchainInfo};
-use sp_consensus::BlockStatus;
-use sp_consensus_aura::sr25519::AuthorityId as AuraId;
-use sp_core::Pair;
-use sp_runtime::{
-	generic::{Era, SignedBlock},
-	traits::{BlakeTwo256, Block as BlockT},
-	Justifications, OpaqueExtrinsic, SaturatedConversion,
-};
-use sp_storage::{ChildInfo, StorageData, StorageKey};
-use std::sync::Arc;
-use substrate_frame_rpc_system::AccountNonceApi;
-pub use tangle_primitives::{AccountId, Balance, Block, BlockNumber, Hash, Header, Index};
-
-/// A set of APIs that polkadot-like runtimes must implement.
-///
-/// This trait has no methods or associated type. It is a concise marker for all the trait bounds
-/// that it contains.
-pub trait RuntimeApiCollection:
-	sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block>
-	+ sp_api::ApiExt<Block>
-	+ sp_block_builder::BlockBuilder<Block>
-	+ substrate_frame_rpc_system::AccountNonceApi<Block, AccountId, Index>
-	+ pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance>
-	+ sp_api::Metadata<Block>
-	+ sp_offchain::OffchainWorkerApi<Block>
-	+ sp_session::SessionKeys<Block>
-	+ fp_rpc::ConvertTransactionRuntimeApi<Block>
-	+ fp_rpc::EthereumRuntimeRPCApi<Block>
-	+ rpc_primitives_debug::DebugRuntimeApi<Block>
-	+ rpc_primitives_txpool::TxPoolRuntimeApi<Block>
-	+ sp_consensus_aura::AuraApi<Block, AuraId>
-	+ sp_consensus_grandpa::GrandpaApi<Block>
-{
-}
-
-impl<Api> RuntimeApiCollection for Api where
-	Api: sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block>
-		+ sp_api::ApiExt<Block>
-		+ sp_block_builder::BlockBuilder<Block>
-		+ substrate_frame_rpc_system::AccountNonceApi<Block, AccountId, Index>
-		+ pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance>
-		+ sp_api::Metadata<Block>
-		+ sp_offchain::OffchainWorkerApi<Block>
-		+ sp_session::SessionKeys<Block>
-		+ fp_rpc::ConvertTransactionRuntimeApi<Block>
-		+ fp_rpc::EthereumRuntimeRPCApi<Block>
-		+ rpc_primitives_debug::DebugRuntimeApi<Block>
-		+ rpc_primitives_txpool::TxPoolRuntimeApi<Block>
-		+ sp_consensus_aura::AuraApi<Block, AuraId>
-		+ sp_consensus_grandpa::GrandpaApi<Block>
-{
-}
-
-/// Config that abstracts over all available client implementations.
-///
-/// For a concrete type there exists [`Client`].
-pub trait AbstractClient<Block, Backend>:
-	BlockchainEvents<Block>
-	+ Sized
-	+ Send
-	+ Sync
-	+ ProvideRuntimeApi<Block>
-	+ HeaderBackend<Block>
-	+ CallApiAt<Block, StateBackend = Backend::State>
-where
-	Block: BlockT,
-	Backend: BackendT<Block>,
-	Backend::State: sp_api::StateBackend<BlakeTwo256>,
-	Self::Api: RuntimeApiCollection,
-{
-}
-
-impl<Block, Backend, Client> AbstractClient<Block, Backend> for Client
-where
-	Block: BlockT,
-	Backend: BackendT<Block>,
-	Backend::State: sp_api::StateBackend<BlakeTwo256>,
-	Client: BlockchainEvents<Block>
-		+ ProvideRuntimeApi<Block>
-		+ HeaderBackend<Block>
-		+ Sized
-		+ Send
-		+ Sync
-		+ CallApiAt<Block, StateBackend = Backend::State>,
-	Client::Api: RuntimeApiCollection,
-{
-}
-
-/// Execute something with the client instance.
-///
-/// As there exist multiple chains inside Tangle, like Tangle itself,
-/// the Testnet, there can exist different kinds of client types. As these
-/// client types differ in the generics that are being used, we can not easily
-/// return them from a function. For returning them from a function there exists
-/// [`Client`]. However, the problem on how to use this client instance still
-/// exists. This trait "solves" it in a dirty way. It requires a type to
-/// implement this trait and than the [`execute_with_client`](ExecuteWithClient:
-/// :execute_with_client) function can be called with any possible client
-/// instance.
-///
-/// In a perfect world, we could make a closure work in this way.
-pub trait ExecuteWithClient {
-	/// The return type when calling this instance.
-	type Output;
-
-	/// Execute whatever should be executed with the given client instance.
-	fn execute_with_client<Client, Api, Backend>(self, client: Arc<Client>) -> Self::Output
-	where
-		Backend: sc_client_api::Backend<Block>,
-		Backend::State: sp_api::StateBackend<BlakeTwo256>,
-		Api: RuntimeApiCollection,
-		Client: AbstractClient<Block, Backend, Api = Api> + 'static;
-}
-
-/// A handle to a tangle client instance.
-///
-/// The tangle service supports multiple different runtimes (Moonbase, tangle
-/// itself, etc). As each runtime has a specialized client, we need to hide them
-/// behind a trait. This is this trait.
-///
-/// When wanting to work with the inner client, you need to use `execute_with`.
-pub trait ClientHandle {
-	/// Execute the given something with the client.
-	fn execute_with<T: ExecuteWithClient>(&self, t: T) -> T::Output;
-}
-
-/// A client instance of tangle.
-#[derive(Clone)]
-pub enum Client {
-	#[cfg(feature = "tangle")]
-	Tangle(Arc<crate::FullClient<tangle_mainnet_runtime::RuntimeApi, crate::TangleExecutor>>),
-	#[cfg(feature = "testnet")]
-	Testnet(Arc<crate::FullClient<tangle_testnet_runtime::RuntimeApi, crate::TestnetExecutor>>),
-}
-
-#[cfg(feature = "tangle")]
-impl From<Arc<crate::FullClient<tangle_mainnet_runtime::RuntimeApi, crate::TangleExecutor>>>
-	for Client
-{
-	fn from(
-		client: Arc<crate::FullClient<tangle_mainnet_runtime::RuntimeApi, crate::TangleExecutor>>,
-	) -> Self {
-		Self::Tangle(client)
-	}
-}
-
-#[cfg(feature = "testnet")]
-impl From<Arc<crate::FullClient<tangle_testnet_runtime::RuntimeApi, crate::TestnetExecutor>>>
-	for Client
-{
-	fn from(
-		client: Arc<crate::FullClient<tangle_testnet_runtime::RuntimeApi, crate::TestnetExecutor>>,
-	) -> Self {
-		Self::Testnet(client)
-	}
-}
-
-impl ClientHandle for Client {
-	fn execute_with<T: ExecuteWithClient>(&self, t: T) -> T::Output {
-		match self {
-			#[cfg(feature = "tangle")]
-			Self::Tangle(client) => T::execute_with_client::<_, _, crate::FullBackend>(t, client.clone()),
-			#[cfg(feature = "testnet")]
-			Self::Testnet(client) => T::execute_with_client::<_, _, crate::FullBackend>(t, client.clone()),
-		}
-	}
-}
-
-macro_rules! match_client {
-	($self:ident, $method:ident($($param:ident),*)) => {
-		match $self {
-			#[cfg(feature = "tangle")]
-			Self::Tangle(client) => client.$method($($param),*),
-			#[cfg(feature = "testnet")]
-			Self::Testnet(client) => client.$method($($param),*),
-		}
-	};
-}
-
-impl sc_client_api::UsageProvider<Block> for Client {
-	fn usage_info(&self) -> sc_client_api::ClientInfo<Block> {
-		match_client!(self, usage_info())
-	}
-}
-
-impl sc_client_api::BlockBackend<Block> for Client {
-	fn block_body(
-		&self,
-		hash: <Block as BlockT>::Hash,
-	) -> sp_blockchain::Result<Option<Vec<<Block as BlockT>::Extrinsic>>> {
-		match_client!(self, block_body(hash))
-	}
-
-	fn block_indexed_body(
-		&self,
-		hash: <Block as BlockT>::Hash,
-	) -> sp_blockchain::Result<Option<Vec<Vec<u8>>>> {
-		match_client!(self, block_indexed_body(hash))
-	}
-
-	fn block(
-		&self,
-		hash: <Block as BlockT>::Hash,
-	) -> sp_blockchain::Result<Option<SignedBlock<Block>>> {
-		match_client!(self, block(hash))
-	}
-
-	fn block_status(&self, hash: <Block as BlockT>::Hash) -> sp_blockchain::Result<BlockStatus> {
-		match_client!(self, block_status(hash))
-	}
-
-	fn justifications(
-		&self,
-		hash: <Block as BlockT>::Hash,
-	) -> sp_blockchain::Result<Option<Justifications>> {
-		match_client!(self, justifications(hash))
-	}
-
-	fn block_hash(
-		&self,
-		number: NumberFor<Block>,
-	) -> sp_blockchain::Result<Option<<Block as BlockT>::Hash>> {
-		match_client!(self, block_hash(number))
-	}
-
-	fn indexed_transaction(
-		&self,
-		hash: <Block as BlockT>::Hash,
-	) -> sp_blockchain::Result<Option<Vec<u8>>> {
-		match_client!(self, indexed_transaction(hash))
-	}
-
-	fn has_indexed_transaction(
-		&self,
-		hash: <Block as BlockT>::Hash,
-	) -> sp_blockchain::Result<bool> {
-		match_client!(self, has_indexed_transaction(hash))
-	}
-
-	fn requires_full_sync(&self) -> bool {
-		match_client!(self, requires_full_sync())
-	}
-}
-
-impl sc_client_api::StorageProvider<Block, crate::FullBackend> for Client {
-	fn storage(
-		&self,
-		hash: <Block as BlockT>::Hash,
-		key: &StorageKey,
-	) -> sp_blockchain::Result<Option<StorageData>> {
-		match_client!(self, storage(hash, key))
-	}
-
-	fn storage_hash(
-		&self,
-		hash: <Block as BlockT>::Hash,
-		key: &StorageKey,
-	) -> sp_blockchain::Result<Option<<Block as BlockT>::Hash>> {
-		match_client!(self, storage_hash(hash, key))
-	}
-
-	fn storage_keys(
-		&self,
-		hash: <Block as BlockT>::Hash,
-		prefix: Option<&StorageKey>,
-		start_key: Option<&StorageKey>,
-	) -> sp_blockchain::Result<
-		KeysIter<<crate::FullBackend as sc_client_api::Backend<Block>>::State, Block>,
-	> {
-		match_client!(self, storage_keys(hash, prefix, start_key))
-	}
-
-	fn storage_pairs(
-		&self,
-		hash: <Block as BlockT>::Hash,
-		key_prefix: Option<&StorageKey>,
-		start_key: Option<&StorageKey>,
-	) -> sp_blockchain::Result<
-		PairsIter<<crate::FullBackend as sc_client_api::Backend<Block>>::State, Block>,
-	> {
-		match_client!(self, storage_pairs(hash, key_prefix, start_key))
-	}
-
-	fn child_storage(
-		&self,
-		hash: <Block as BlockT>::Hash,
-		child_info: &ChildInfo,
-		key: &StorageKey,
-	) -> sp_blockchain::Result<Option<StorageData>> {
-		match_client!(self, child_storage(hash, child_info, key))
-	}
-
-	fn child_storage_keys(
-		&self,
-		hash: <Block as BlockT>::Hash,
-		child_info: ChildInfo,
-		prefix: Option<&StorageKey>,
-		start_key: Option<&StorageKey>,
-	) -> sp_blockchain::Result<
-		KeysIter<<crate::FullBackend as sc_client_api::Backend<Block>>::State, Block>,
-	> {
-		match_client!(self, child_storage_keys(hash, child_info, prefix, start_key))
-	}
-
-	fn child_storage_hash(
-		&self,
-		hash: <Block as BlockT>::Hash,
-		child_info: &ChildInfo,
-		key: &StorageKey,
-	) -> sp_blockchain::Result<Option<<Block as BlockT>::Hash>> {
-		match_client!(self, child_storage_hash(hash, child_info, key))
-	}
-}
-
-impl sp_blockchain::HeaderBackend<Block> for Client {
-	fn header(&self, hash: Hash) -> sp_blockchain::Result<Option<Header>> {
-		match_client!(self, header(hash))
-	}
-
-	fn info(&self) -> sp_blockchain::Info<Block> {
-		match_client!(self, info())
-	}
-
-	fn status(&self, hash: Hash) -> sp_blockchain::Result<sp_blockchain::BlockStatus> {
-		match_client!(self, status(hash))
-	}
-
-	fn number(&self, hash: Hash) -> sp_blockchain::Result<Option<BlockNumber>> {
-		match_client!(self, number(hash))
-	}
-
-	fn hash(&self, number: NumberFor<Block>) -> sp_blockchain::Result<Option<Hash>> {
-		match_client!(self, hash(number))
-	}
-}
-
-/// A client instance of tangle.
-#[derive(Clone)]
-pub enum ClientUncheckedExtrinsic {
-	#[cfg(feature = "tangle")]
-	Tangle(Arc<tangle_mainnet_runtime::UncheckedExtrinsic>),
-	#[cfg(feature = "testnet")]
-	Testnet(Arc<tangle_testnet_runtime::UncheckedExtrinsic>),
-}
-
-impl Into<OpaqueExtrinsic> for ClientUncheckedExtrinsic {
-	fn into(self) -> OpaqueExtrinsic {
-		match self {
-			#[cfg(feature = "tangle")]
-			ClientUncheckedExtrinsic::Tangle(extrinsic) =>
-				OpaqueExtrinsic::from_bytes(&extrinsic.encode()).unwrap(),
-			#[cfg(feature = "testnet")]
-			ClientUncheckedExtrinsic::Testnet(extrinsic) =>
-				OpaqueExtrinsic::from_bytes(&extrinsic.encode()).unwrap(),
-		}
-	}
-}
-
-impl Client {
-	pub fn chain_info(&self) -> BlockchainInfo<Block> {
-		match self {
-			#[cfg(feature = "tangle")]
-			Self::Tangle(client) => client.chain_info(),
-			#[cfg(feature = "testnet")]
-			Self::Testnet(client) => client.chain_info(),
-		}
-	}
-
-	pub fn fetch_nonce(&self, account: sp_core::sr25519::Pair) -> u32 {
-		let best_hash = self.chain_info().best_hash;
-		macro_rules! fetch_nonce_from_client {
-			($client:expr) => {
-				$client
-					.runtime_api()
-					.account_nonce(best_hash, account.public().into())
-					.expect("Fetching account nonce works; qed")
-			};
-		}
-
-		match self {
-			#[cfg(feature = "tangle")]
-			Self::Tangle(client) => fetch_nonce_from_client!(client),
-			#[cfg(feature = "testnet")]
-			Self::Testnet(client) => fetch_nonce_from_client!(client),
-		}
-	}
-
-	/// Create a transaction using the given `call`
-	///
-	/// The transaction will be signed by `sender`. If `nonce` is `None` it will be fetched from the
-	/// state of the best block.
-	///
-	/// Note: Should only be used for tests.
-	pub fn create_extrinsic(
-		&self,
-		sender: sp_core::sr25519::Pair,
-		#[cfg(feature = "tangle")] function: impl Into<tangle_mainnet_runtime::RuntimeCall>,
-		#[cfg(feature = "testnet")] function: impl Into<tangle_testnet_runtime::RuntimeCall>,
-		nonce: Option<u32>,
-	) -> ClientUncheckedExtrinsic {
-		let function = function.into();
-		let genesis_hash = self.block_hash(0).ok().flatten().expect("Genesis block exists; qed");
-		let best_hash = self.chain_info().best_hash;
-		let best_block = self.chain_info().best_number;
-		let nonce = nonce.unwrap_or_else(|| self.fetch_nonce(sender.clone()));
-
-		macro_rules! create_extrinsic {
-			($runtime:ident) => {{
-				let period = $runtime::BlockHashCount::get()
-					.checked_next_power_of_two()
-					.map(|c| c / 2)
-					.unwrap_or(2) as u64;
-				let tip = 0;
-				let extra: $runtime::SignedExtra = (
-					frame_system::CheckNonZeroSender::<$runtime::Runtime>::new(),
-					frame_system::CheckSpecVersion::<$runtime::Runtime>::new(),
-					frame_system::CheckTxVersion::<$runtime::Runtime>::new(),
-					frame_system::CheckGenesis::<$runtime::Runtime>::new(),
-					frame_system::CheckEra::<$runtime::Runtime>::from(Era::Mortal(
-						period,
-						best_block.saturated_into(),
-					)),
-					frame_system::CheckNonce::<$runtime::Runtime>::from(nonce),
-					frame_system::CheckWeight::<$runtime::Runtime>::new(),
-					pallet_transaction_payment::ChargeTransactionPayment::<$runtime::Runtime>::from(
-						tip,
-					),
-				);
-
-				let raw_payload = $runtime::SignedPayload::from_raw(
-					function.clone(),
-					extra.clone(),
-					(
-						(),
-						$runtime::VERSION.spec_version,
-						$runtime::VERSION.transaction_version,
-						genesis_hash,
-						best_hash,
-						(),
-						(),
-						(),
-					),
-				);
-				let signature = raw_payload.using_encoded(|e| sender.sign(e));
-
-				$runtime::UncheckedExtrinsic::new_signed(
-					function,
-					sp_runtime::AccountId32::from(sender.public()).into(),
-					$runtime::Signature::Sr25519(signature),
-					extra,
-				)
-			}};
-		}
-
-		#[cfg(feature = "tangle")]
-		let extrinsic =
-			ClientUncheckedExtrinsic::Tangle(Arc::new(create_extrinsic!(tangle_mainnet_runtime)));
-		#[cfg(feature = "testnet")]
-		let extrinsic =
-			ClientUncheckedExtrinsic::Testnet(Arc::new(create_extrinsic!(tangle_testnet_runtime)));
-
-		ClientUncheckedExtrinsic::from(extrinsic)
-	}
-}
