--- node/src/benchmarking.rs
@@ -1,121 +1,0 @@
-// This file is part of Substrate.
-
-// Copyright (C) 2022 Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-//! Setup code for [`super::command`] which would otherwise bloat that module.
-//!
-//! Should only be used for benchmarking as it may break in other contexts.
-
-use tangle_service::client::Client;
-
-use sc_cli::Result;
-use sp_inherents::{InherentData, InherentDataProvider};
-use sp_keyring::Sr25519Keyring;
-use sp_runtime::OpaqueExtrinsic;
-use std::{sync::Arc, time::Duration};
-#[cfg(feature = "tangle")]
-use tangle_mainnet_runtime::{AccountId, Balance, BalancesCall, SystemCall};
-#[cfg(feature = "testnet")]
-use tangle_testnet_runtime::{AccountId, Balance, BalancesCall, SystemCall};
-
-/// Generates `System::Remark` extrinsics for the benchmarks.
-///
-/// Note: Should only be used for benchmarking.
-pub struct RemarkBuilder {
-	client: Arc<Client>,
-}
-
-impl RemarkBuilder {
-	/// Creates a new [`Self`] from the given client.
-	pub fn new(client: Arc<Client>) -> Self {
-		Self { client }
-	}
-}
-
-impl frame_benchmarking_cli::ExtrinsicBuilder for RemarkBuilder {
-	fn pallet(&self) -> &str {
-		"system"
-	}
-
-	fn extrinsic(&self) -> &str {
-		"remark"
-	}
-
-	fn build(&self, nonce: u32) -> std::result::Result<OpaqueExtrinsic, &'static str> {
-		let acc = Sr25519Keyring::Bob.pair();
-		let extrinsic: OpaqueExtrinsic = self
-			.client
-			.create_extrinsic(acc, SystemCall::remark { remark: vec![] }, Some(nonce))
-			.into();
-
-		Ok(extrinsic)
-	}
-}
-
-/// Generates `Balances::TransferKeepAlive` extrinsics for the benchmarks.
-///
-/// Note: Should only be used for benchmarking.
-pub struct TransferKeepAliveBuilder {
-	client: Arc<Client>,
-	dest: AccountId,
-	value: Balance,
-}
-
-impl TransferKeepAliveBuilder {
-	/// Creates a new [`Self`] from the given client.
-	pub fn new(client: Arc<Client>, dest: AccountId, value: Balance) -> Self {
-		Self { client, dest, value }
-	}
-}
-
-impl frame_benchmarking_cli::ExtrinsicBuilder for TransferKeepAliveBuilder {
-	fn pallet(&self) -> &str {
-		"balances"
-	}
-
-	fn extrinsic(&self) -> &str {
-		"transfer_keep_alive"
-	}
-
-	fn build(&self, nonce: u32) -> std::result::Result<OpaqueExtrinsic, &'static str> {
-		let acc = Sr25519Keyring::Bob.pair();
-		let extrinsic: OpaqueExtrinsic = self
-			.client
-			.create_extrinsic(
-				acc,
-				BalancesCall::transfer_keep_alive {
-					dest: self.dest.clone().into(),
-					value: self.value,
-				},
-				Some(nonce),
-			)
-			.into();
-
-		Ok(extrinsic)
-	}
-}
-
-/// Generates inherent data for the `benchmark overhead` command.
-pub fn inherent_benchmark_data() -> Result<InherentData> {
-	let mut inherent_data = InherentData::new();
-	let d = Duration::from_millis(0);
-	let timestamp = sp_timestamp::InherentDataProvider::new(d.into());
-
-	futures::executor::block_on(timestamp.provide_inherent_data(&mut inherent_data))
-		.map_err(|e| format!("creating inherent data: {e}"))?;
-	Ok(inherent_data)
-}
