+++ node/src/distributions/data/edgeware_genesis_participants.json
