+++ node/src/distributions/mainnet.rs
@@ -30,19 +30,17 @@ fn read_contents_to_substrate_accounts(path_str: &str) -> BTreeMap<AccountId, f6
 }
 
 fn get_edgeware_genesis_list() -> Vec<H160> {
-	read_contents_to_evm_accounts(
-		"node/service/src/distributions/data/edgeware_genesis_participants.json",
-	)
+	read_contents_to_evm_accounts("node/src/distributions/data/edgeware_genesis_participants.json")
 }
 
 fn get_edgeware_snapshot_list() -> BTreeMap<AccountId32, f64> {
 	read_contents_to_substrate_accounts(
-		"node/service/src/distributions/data/edgeware_snapshot_distribution.json",
+		"node/src/distributions/data/edgeware_snapshot_distribution.json",
 	)
 }
 
 fn get_discord_list() -> Vec<H160> {
-	read_contents_to_evm_accounts("node/service/src/distributions/data/discord_evm_addresses.json")
+	read_contents_to_evm_accounts("node/src/distributions/data/discord_evm_addresses.json")
 }
 
 pub fn get_edgeware_genesis_balance_distribution() -> Vec<(H160, GenesisAccount)> {
