+++ node/src/distributions/testnet.rs
@@ -61,19 +61,17 @@ fn read_contents_to_substrate_accounts(path_str: &str) -> Vec<AccountId> {
 }
 
 fn get_edgeware_genesis_list() -> Vec<H160> {
-	read_contents_to_evm_accounts(
-		"node/service/src/distributions/data/edgeware_genesis_participants.json",
-	)
+	read_contents_to_evm_accounts("node/src/distributions/data/edgeware_genesis_participants.json")
 }
 
 fn get_edgeware_snapshot_list() -> Vec<AccountId> {
 	read_contents_to_substrate_accounts(
-		"node/service/src/distributions/data/edgeware_snapshot_participants.json",
+		"node/src/distributions/data/edgeware_snapshot_participants.json",
 	)
 }
 
 fn get_discord_list() -> Vec<H160> {
-	read_contents_to_evm_accounts("node/service/src/distributions/data/discord_evm_addresses.json")
+	read_contents_to_evm_accounts("node/src/distributions/data/discord_evm_addresses.json")
 }
 
 pub fn get_evm_balance_distribution() -> Vec<(H160, GenesisAccount)> {
