+++ node/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod chainspec;
+pub mod cli;
+pub mod distributions;
+pub mod eth;
+pub mod mainnet_fixtures;
+pub mod rpc;
+pub mod service;
+pub mod testnet_fixtures;
