+++ primitives/Cargo.toml
@@ -15,7 +15,6 @@ scale-info = { workspace = true }
 serde = { workspace = true }
 smallvec = { workspace = true }
 sp-arithmetic = { workspace = true }
-sp-consensus-aura = { workspace = true }
 sp-consensus-babe = { workspace = true }
 sp-core = { workspace = true }
 sp-runtime = { workspace = true }
@@ -41,7 +40,6 @@ std = [
   "frame-support/std",
   "sp-runtime/std",
   "sp-core/std",
-  "sp-consensus-aura/std",
   "ark-std?/std",
 ]
 verifying = [
